Update message component

+++ client/src/app/messages/messages.component.ts
@@ -3,6 +3,7 @@ import {Message} from "../_models/message";
 import {Pagination} from "../_models/pagination";
 import {MessageParams} from "../_models/messageParams";
 import {MessageService} from "../_services/message.service";
+import {ConfirmService} from "../_services/confirm.service";
 
 @Component({
   selector: 'app-messages',
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   messageParams = new MessageParams();
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,9 +34,13 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id == id), 1);
-    })
+    this.confirmService.confirm('Confirm delete message', 'this cannot be undone').subscribe(result => {
+      if(result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+        this.messages.splice(this.messages.findIndex(m => m.id == id), 1);
+      })
+    }});
+
   }
 
   pageChanged(event: any) {
